requirements.txt
streamlit
python-dotenv
langchain
openai
youtube-transcript-api
faiss-cpu
tiktoken